@@ -1,123 +1,125 @@
 {
-   "name": "@stratio/egeo",
-   "version": "",
-   "description": "Egeo is an angular common components library",
-   "author": "Stratio",
-   "scripts": {
-      "start": "ng serve",
-      "release": "gulp publish",
-      "clean": "gulp clean",
-      "build": "gulp aot:build",
-      "build:web:aot": "gulp aot:build",
-      "build:lib": "gulp egeo:build-release:clean",
-      "build:theme": "gulp build:styles",
-      "build:demo-app": "npm run clean && npm run demo-app:create-dist && npm run demo-app:generate",
-      "demo-app:generate": "ng build --prod --aot=false --progress=false && gulp demo-app:replace:assets",
-      "demo-app:create-dist": "gulp demo-app:create-dist",
-      "test": "karma start test/karma.conf.js",
-      "test:watch": "karma start test/karma.conf.js --auto-watch --no-single-run",
-      "lint": "gulp lint",
-      "replace": "json -I -f package.json -e",
-      "commit": "git-cz",
-      "serve:coverage": "http-server target/html",
-      "serve:build": "npm run build:demo-ap && http-server dist/demo-app"
-   },
-   "repository": {
-      "type": "git",
-      "url": "https://github.com/Stratio/egeo"
-   },
-   "license": "Apache-2.0",
-   "keywords": [
-      "angular",
-      "components",
-      "egeo",
-      "rollup",
-      "karma",
-      "gulp"
-   ],
-   "dependencies": {
-      "angular2-virtual-scroll": "0.3.1",
-      "lodash": "4.17.4",
-      "reflect-metadata": "0.1.10",
-      "tslib": "1.7.1"
-   },
-   "devDependencies": {
-      "@angular/animations": "4.4.3",
-      "@angular/cli": "1.6.5",
-      "@angular/common": "4.4.3",
-      "@angular/compiler": "4.4.3",
-      "@angular/compiler-cli": "4.4.3",
-      "@angular/core": "4.4.3",
-      "@angular/forms": "4.4.3",
-      "@angular/http": "4.4.3",
-      "@angular/platform-browser": "4.4.3",
-      "@angular/platform-browser-dynamic": "4.4.3",
-      "@angular/router": "4.4.3",
-      "@stratio/egeo-ui-base": "1.4.1",
-      "@types/chalk": "0.4.31",
-      "@types/fs-extra": "3.0.3",
-      "@types/glob": "5.0.30",
-      "@types/gulp": "3.8.32",
-      "@types/jasmine": "2.8.6",
-      "@types/js-beautify": "0.0.30",
-      "@types/lodash": "4.14.68",
-      "@types/minimist": "1.2.0",
-      "@types/node": "6.0.60",
-      "@types/run-sequence": "0.0.29",
-      "@types/showdown": "1.4.32",
-      "autoprefixer": "7.1.1",
-      "chalk": "2.1.0",
-      "commitizen": "2.9.6",
-      "core-js": "2.4.1",
-      "cssnano": "3.10.0",
-      "fs-extra": "3.0.1",
-      "gulp": "3.9.1",
-      "gulp-clean": "0.3.2",
-      "gulp-clean-css": "3.5.0",
-      "gulp-connect": "5.0.0",
-      "gulp-htmlmin": "3.0.0",
-      "gulp-if": "2.0.2",
-      "gulp-postcss": "7.0.0",
-      "gulp-remove-html-comments": "1.0.1",
-      "gulp-sass": "3.1.0",
-      "gulp-sourcemaps": "2.6.0",
-      "http-server": "0.10.0",
-      "intl": "1.2.5",
-      "jasmine-core": "2.9.1",
-      "js-beautify": "1.6.14",
-      "json": "9.0.6",
-      "karma": "2.0.0",
-      "karma-jasmine": "1.1.1",
-      "karma-junit-reporter": "1.2.0",
-      "karma-mocha-reporter": "2.2.5",
-      "karma-phantomjs-launcher": "1.0.4",
-      "karma-typescript": "3.0.12",
-      "karma-typescript-angular2-transform": "1.0.2",
-      "madge": "1.6.0",
-      "markdown-to-html-pipe": "1.2.4",
-      "minimist": "1.2.0",
-      "resolve-bin": "0.4.0",
-      "rollup": "0.52.0",
-      "rollup-plugin-bundle-size": "1.0.1",
-      "rollup-plugin-node-resolve": "3.0.0",
-      "rollup-plugin-strip-banner": "0.2.0",
-      "run-sequence": "2.2.0",
-      "rxjs": "5.5.2",
-      "scss-bundle": "2.1.0",
-      "showdown": "1.8.2",
-      "sorcery": "0.10.0",
-      "stylelint": "8.3.0",
-      "ts-node": "3.2.0",
-      "tsconfig-paths": "2.2.0",
-      "tslint": "5.5.0",
-      "typescript": "2.3.4",
-      "uglify-js": "2.8.22",
-      "word-wrap": "1.2.3",
-      "zone.js": "0.8.18"
-   },
-   "config": {
-      "commitizen": {
-         "path": "./tools/commitizen-format/st-commitizen.js"
-      }
-   }
+  "name": "@stratio/egeo",
+  "version": "",
+  "description": "Egeo is an angular common components library",
+  "author": "Stratio",
+  "scripts": {
+    "start": "ng serve",
+    "release": "gulp publish",
+    "clean": "gulp clean",
+    "build": "gulp aot:build",
+    "build:web:aot": "gulp aot:build",
+    "build:lib": "gulp egeo:build-release:clean",
+    "build:theme": "gulp build:styles",
+    "build:demo-app": "npm run clean && npm run demo-app:create-dist && npm run demo-app:generate",
+    "demo-app:generate": "ng build --prod --aot=false --progress=false && gulp demo-app:replace:assets",
+    "demo-app:create-dist": "gulp demo-app:create-dist",
+    "test": "karma start test/karma.conf.js",
+    "test:watch": "karma start test/karma.conf.js --auto-watch --no-single-run",
+    "lint": "gulp lint",
+    "replace": "json -I -f package.json -e",
+    "commit": "git-cz",
+    "serve:coverage": "http-server target/html",
+    "serve:build": "npm run build:demo-ap && http-server dist/demo-app"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Stratio/egeo"
+  },
+  "license": "Apache-2.0",
+  "keywords": [
+    "angular",
+    "components",
+    "egeo",
+    "rollup",
+    "karma",
+    "gulp"
+  ],
+  "dependencies": {
+    "@angular/cli": "^6.0.7",
+    "angular2-virtual-scroll": "0.3.1",
+    "lodash": "4.17.4",
+    "reflect-metadata": "0.1.10",
+    "tslib": "1.7.1"
+  },
+  "devDependencies": {
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@angular/tsc-wrapped": "4.4.6",
+    "@stratio/egeo-ui-base": "1.4.1",
+    "@types/chalk": "0.4.31",
+    "@types/fs-extra": "3.0.3",
+    "@types/glob": "5.0.30",
+    "@types/gulp": "3.8.32",
+    "@types/jasmine": "2.8.6",
+    "@types/js-beautify": "0.0.30",
+    "@types/lodash": "4.14.109",
+    "@types/minimist": "1.2.0",
+    "@types/node": "6.0.60",
+    "@types/run-sequence": "0.0.29",
+    "@types/showdown": "1.4.32",
+    "autoprefixer": "7.1.1",
+    "chalk": "2.1.0",
+    "commitizen": "2.9.6",
+    "core-js": "2.4.1",
+    "cssnano": "3.10.0",
+    "fs-extra": "3.0.1",
+    "gulp": "3.9.1",
+    "gulp-clean": "0.3.2",
+    "gulp-clean-css": "3.5.0",
+    "gulp-connect": "5.0.0",
+    "gulp-htmlmin": "3.0.0",
+    "gulp-if": "2.0.2",
+    "gulp-postcss": "7.0.0",
+    "gulp-remove-html-comments": "1.0.1",
+    "gulp-sass": "3.1.0",
+    "gulp-sourcemaps": "2.6.0",
+    "http-server": "0.10.0",
+    "intl": "1.2.5",
+    "jasmine-core": "2.9.1",
+    "js-beautify": "1.6.14",
+    "json": "9.0.6",
+    "karma": "2.0.0",
+    "karma-jasmine": "1.1.1",
+    "karma-junit-reporter": "1.2.0",
+    "karma-mocha-reporter": "2.2.5",
+    "karma-phantomjs-launcher": "1.0.4",
+    "karma-typescript": "3.0.12",
+    "karma-typescript-angular2-transform": "1.0.2",
+    "madge": "1.6.0",
+    "markdown-to-html-pipe": "1.2.4",
+    "minimist": "1.2.0",
+    "resolve-bin": "0.4.0",
+    "rollup": "0.52.0",
+    "rollup-plugin-bundle-size": "1.0.1",
+    "rollup-plugin-node-resolve": "3.0.0",
+    "rollup-plugin-strip-banner": "0.2.0",
+    "run-sequence": "2.2.0",
+    "rxjs": "6.2.0",
+    "scss-bundle": "2.1.0",
+    "showdown": "1.8.2",
+    "sorcery": "0.10.0",
+    "stylelint": "8.3.0",
+    "ts-node": "3.2.0",
+    "tsconfig-paths": "2.2.0",
+    "tslint": "5.5.0",
+    "typescript": "2.7.2",
+    "uglify-js": "2.8.22",
+    "word-wrap": "1.2.3",
+    "zone.js": "0.8.26",
+    "@angular-devkit/build-angular": "~0.6.6"
+  },
+  "config": {
+    "commitizen": {
+      "path": "./tools/commitizen-format/st-commitizen.js"
+    }
+  }
 }
