@@ -9,8 +9,8 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { Observable } from 'rxjs';
+import { BehaviorSubject } from 'rxjs';
 
 import { StDemoLogger, StDemoLoggerSeverity } from './st-demo-loger.model';
 
