@@ -11,8 +11,8 @@
 import { DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed, fakeAsync, tick, discardPeriodicTasks } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable } from 'rxjs';
+import { Observer } from 'rxjs';
 
 import { StAlertsComponent } from './st-alerts.component';
 import { StAlertsService } from './st-alerts.service';
