@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 export interface EgeoResolverKeys {
    path: string;
