@@ -9,8 +9,8 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { cloneDeep as _cloneDeep } from 'lodash';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable } from 'rxjs';
+import { Observer } from 'rxjs';
 import { EgeoResolverKeys, TranslateableElement, TranslateServiceType } from './egeo-resolve-model';
 import { EgeoResolveService } from './egeo-resolve.service';
 
