@@ -11,7 +11,7 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, forwardRef, Input,
    OnChanges, OnDestroy, OnInit, ViewChildren } from '@angular/core';
 import { ControlValueAccessor, FormControl, NG_VALIDATORS, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { StTextareaError } from './st-textarea.error.model';
 
