@@ -34,33 +34,31 @@ const ROLLUP_GLOBALS = {
    '@stratio/egeo': 'egeo',
 
    // Rxjs dependencies
-   'rxjs/BehaviorSubject': 'Rx',
-   'rxjs/Observable': 'Rx',
-   'rxjs/Subject': 'Rx',
-   'rxjs/Subscription': 'Rx',
-   'rxjs/add/observable/combineLatest': 'Rx.Observable',
-   'rxjs/add/observable/forkJoin': 'Rx.Observable',
-   'rxjs/add/observable/fromEvent': 'Rx.Observable',
-   'rxjs/add/observable/merge': 'Rx.Observable',
-   'rxjs/add/observable/of': 'Rx.Observable',
-   'rxjs/add/observable/throw': 'Rx.Observable',
-   'rxjs/add/operator/auditTime': 'Rx.Observable.prototype',
-   'rxjs/add/operator/catch': 'Rx.Observable.prototype',
-   'rxjs/add/operator/debounceTime': 'Rx.Observable.prototype',
-   'rxjs/add/operator/do': 'Rx.Observable.prototype',
-   'rxjs/add/operator/filter': 'Rx.Observable.prototype',
-   'rxjs/add/operator/finally': 'Rx.Observable.prototype',
-   'rxjs/add/operator/first': 'Rx.Observable.prototype',
-   'rxjs/add/operator/let': 'Rx.Observable.prototype',
-   'rxjs/add/operator/map': 'Rx.Observable.prototype',
-   'rxjs/add/operator/share': 'Rx.Observable.prototype',
-   'rxjs/add/operator/startWith': 'Rx.Observable.prototype',
-   'rxjs/add/operator/switchMap': 'Rx.Observable.prototype',
-   'rxjs/add/operator/takeUntil': 'Rx.Observable.prototype',
-   'rxjs/add/operator/toPromise': 'Rx.Observable.prototype',
+   'rxjs': 'Rx',
+   'rxjs/operators': 'Rx.Observable.prototype',
+   // 'rxjs/add/observable/combineLatest': 'Rx.Observable',
+   // 'rxjs/add/observable/forkJoin': 'Rx.Observable',
+   // 'rxjs/add/observable/fromEvent': 'Rx.Observable',
+   // 'rxjs/add/observable/merge': 'Rx.Observable',
+   // 'rxjs/add/observable/of': 'Rx.Observable',
+   // 'rxjs/add/observable/throw': 'Rx.Observable',
+   // 'rxjs/add/operator/auditTime': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/catch': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/debounceTime': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/do': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/filter': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/finally': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/first': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/let': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/map': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/share': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/startWith': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/switchMap': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/takeUntil': 'Rx.Observable.prototype',
+   // 'rxjs/add/operator/toPromise': 'Rx.Observable.prototype',
 
    // Other dependencies
-   'lodash/index': 'lodash',
+   'lodash': 'lodash',
    'angular2-virtual-scroll': 'angular2-virtual-scroll'
 };
 
